requirements.txt
selenium~=4.20.0
PyYAML~=6.0.1
